requirements.txt
# Core Framework
langgraph==0.2.28
langchain==0.3.1
langchain-core==0.3.6
langchain-community==0.3.1

# OCI SDK and AI
oci==2.129.1
oracle-ads==2.11.9

# Database
oracledb==2.4.1
cx-Oracle==8.3.0

# Vector Store
chromadb==0.5.5
faiss-cpu==1.8.0

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
gradio==4.44.0

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0

# Infrastructure as Code
python-terraform==0.10.1

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
aiofiles==24.1.0

# Observability
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
python-json-logger==2.0.7

# HTTP and Async
httpx==0.27.2
aiohttp==3.10.8
requests==2.32.3

# Validation
jsonschema==4.23.0

# CLI
click==8.1.7
rich==13.8.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# MCP Protocol
mcp==0.9.0

# Diagram Generation
diagrams==0.24.1
matplotlib==3.9.2
networkx==3.4.1
graphviz==0.20.3

# Embeddings and LLM
cohere==5.11.0
tiktoken==0.8.0

# Security
cryptography==43.0.1
